server/requirements.txt
asn1crypto (0.24.0)
cffi (1.11.4)
click (6.7)
cryptography (2.1.4)
Flask (0.12.2)
idna (2.6)
itsdangerous (0.24)
Jinja2 (2.10)
MarkupSafe (1.0)
pip (9.0.1)
pkg-resources (0.0.0)
pycparser (2.18)
pycryptodome (3.4.7)
PyJWT (1.5.3)
setuptools (20.7.0)
six (1.11.0)
Werkzeug (0.13)
